move all the way to _make.sh

+++ build/_make.sh
@@ -8,7 +8,7 @@ repo () {
 	python build/tweaks/fork.py
 }
 repo_escaped () {
-	echo "$(repo)" | sed -e "s/\//\\\\\//g"
+	repo | sed -e "s/\//\\\\\//g"
 }
 slug () {
 	cat build/slug.txt
@@ -18,12 +18,16 @@ should_create_upstream () {
 }
 
 quiet () {
-	if [ $# -eq 0 ] ; then
-		echo "Missing arguments"
+	if [ $# -ne 1 ] ; then
+		echo "Missing command"
 		return 1
 	fi
-
-	$@ > $TMP 2>&1 || cat $TMP
+	output=$(echo "$1" | xargs git > $TMP 2>&1)
+	if [ $output ]; then
+		if [ cat $TMP | grep "no changes" | wc -l -eq 0 ]; then
+			cat $TMP
+		fi
+	fi
 	rm $TMP
 }
 
@@ -33,16 +37,18 @@ if [ $# -ne 1 ] ; then
 fi
 
 if [ "$1" = "init" ] ; then
-	printf "Please enter the name of your curriculum: "
-	read TITLE
-	echo $(title) > build/title.txt
-	echo "Filling in templates..."
-	sed -e "s/{{ title }}/$(title)/g" -e "s/{{ repo }}/$(repo)_escaped/g" "build/template.md" > "$(title).md"
-	echo "Syncing build/title.txt with $(repo)"
-	quiet git reset HEAD .
-	quiet git add build/title.txt
-	quiet git commit -m "Initialize repo"
-	quiet git push -u origin master
+	if [ ! -f "build/title.txt" ]; then
+		printf "Please enter the name of your curriculum: "
+		read TITLE
+		echo $TITLE > build/title.txt
+		echo "Filling in templates..."
+		sed -e "s/{{ title }}/$TITLE/g" -e "s/{{ repo }}/$(repo_escaped)/g" "build/template.md" > "$TITLE.md"
+		echo "Syncing build/title.txt with $(repo)"
+		quiet 'reset HEAD .'
+		quiet 'add build/title.txt'
+		quiet 'commit -m "Initialize repo"'
+		quiet 'push -u origin master'
+	fi
 elif [ "$1" = "fix-remotes" ] ; then
 	if [ -z "$(should_create_upstream)" ] ; then
 	echo "Make a new repo on GitHub, and find it's SSH / HTTPS url, and paste it below."
@@ -52,24 +58,27 @@ elif [ "$1" = "fix-remotes" ] ; then
 	quiet git remote add origin $REMOTE
 	fi
 elif [ "$1" = "build" ] ; then
+	TITLE_TMP="$(title)"
 	cd build
-	cp "../$(title).md" .
-	cp template.css "$(title).css"
-	python md2html.py "$(title).md" ;\
-	mv "$(title).html" ../output.html
-	rm -f "$(title).css" "$(title).md"
+	cp "../$TITLE_TMP.md" .
+	cp template.css "$TITLE_TMP.css"
+	python md2html.py "$TITLE_TMP.md"
+	mv "$TITLE_TMP.html" ../output.html
+	rm -f "$TITLE_TMP.css" "$TITLE_TMP.md"
 elif [ "$1" = "pre-deploy" ] ; then
-	echo "Enter a path slug for your curriculum: \"$(title)\".  It should contain only lowercase letters, numbers, and dashes."
-	echo "Examples: python, javascript, advanced-java\n"
-	printf "Path: "
-	read SLUG
-	echo $SLUG > build/slug.txt
-	ssh adi-website "mkdir -p /srv/learn/public_html/$SLUG"
-	echo "Syncing build/slug.txt with $(repo)"
-	quiet git reset HEAD .
-	quiet git add build/slug.txt
-	quiet git commit -m "Setup Deploy"
-	quiet git push -u origin master
+	if [ ! -f "build/slug.txt" ]; then
+		echo "Enter a path slug for your curriculum: \"$(title)\".  It should contain only lowercase letters, numbers, and dashes."
+		echo "Examples: python, javascript, advanced-java\n"
+		printf "Path: "
+		read SLUG
+		echo $SLUG > build/slug.txt
+		ssh adi-website "mkdir -p /srv/learn/public_html/$SLUG"
+		echo "Syncing build/slug.txt with $(repo)"
+		quiet 'reset HEAD .'
+		quiet 'add build/slug.txt'
+		quiet 'commit -m "Setup Deploy"'
+		quiet 'push -u origin master'
+	fi
 elif [ "$1" = "deploy" ] ; then
 	if scp output.html adi-website:/srv/learn/public_html/$(slug)/index.html > /dev/null; then
 		echo "Deployed to http://learn.adicu.com/$(slug)"
